Rearrange few views and move heroku version to /admin
