+++ open_event/views/admin/home.py
@@ -17,11 +17,21 @@ from open_event.helpers.oauth import OAuth, FbOAuth
 def intended_url():
     return request.args.get('next') or url_for('.index')
 
+def chunks(l, n):
+    """Yield successive n-sized chunks from l."""
+    for i in xrange(0, len(l), n):
+        yield l[i:i + n]
 
 class MyHomeView(AdminIndexView):
+
     @expose('/')
     def index(self):
-        return self.render('gentelella/admin/index.html')
+        call_for_papers_evs = DataGetter.get_call_for_speakers_events().all()
+        published_events = DataGetter.get_all_published_events().all()
+        # Provide data as chunks of 6 for the UI to render in a proper way
+        return self.render('gentelella/index.html',
+                           call_for_papers_evs=list(chunks(call_for_papers_evs, 6)),
+                           published_events=list(chunks(published_events, 6)))
 
     @expose('/login/', methods=('GET', 'POST'))
     def login_view(self):
